Minor fixes

+++ rust/src/day01.rs
@@ -1,8 +1,8 @@
 use std::convert::TryInto;
 use std::iter::zip;
 
-pub fn part_a(input: &str) -> i32 {
-    let depths: Vec<i32> = input.lines().map(|s| s.parse().unwrap()).collect();
+pub fn part_a(input: &str) -> i64 {
+    let depths: Vec<i64> = input.lines().map(|s| s.parse().unwrap()).collect();
     zip(depths.iter(), depths.iter().skip(1))
         .filter(|(x, y)| x < y)
         .count()
@@ -10,8 +10,8 @@ pub fn part_a(input: &str) -> i32 {
         .unwrap()
 }
 
-pub fn part_b(input: &str) -> i32 {
-    let depths: Vec<i32> = input.lines().map(|s| s.parse().unwrap()).collect();
+pub fn part_b(input: &str) -> i64 {
+    let depths: Vec<i64> = input.lines().map(|s| s.parse().unwrap()).collect();
     zip(depths.iter(), depths.iter().skip(3))
         .filter(|(x, y)| x < y)
         .count()
