+++ rust/src/day03.rs
@@ -1,7 +1,7 @@
-pub fn part_a(input: &str) -> i32 {
+pub fn part_a(input: &str) -> i64 {
     let lines: Vec<&str> = input.lines().collect();
     let len = lines[0].len();
-    let gamma: i32 = (0..len)
+    let gamma: i64 = (0..len)
         .map(|i| {
             let count = lines.iter().fold(0, |acc, line| match line.as_bytes()[i] {
                 b'0' => acc - 1,
@@ -19,7 +19,7 @@ pub fn part_a(input: &str) -> i32 {
     gamma * epsilon
 }
 
-fn find_rating(mut lines: Vec<&str>, use_most_common: bool) -> i32 {
+fn find_rating(mut lines: Vec<&str>, use_most_common: bool) -> i64 {
     for i in 0..lines[0].len() {
         let count = lines.iter().fold(0, |acc, line| match line.as_bytes()[i] {
             b'0' => acc - 1,
@@ -37,13 +37,13 @@ fn find_rating(mut lines: Vec<&str>, use_most_common: bool) -> i32 {
             .cloned()
             .collect();
         if lines.len() == 1 {
-            return i32::from_str_radix(lines[0], 2).unwrap();
+            return i64::from_str_radix(lines[0], 2).unwrap();
         }
     }
     panic!("failed to calculate rating");
 }
 
-pub fn part_b(input: &str) -> i32 {
+pub fn part_b(input: &str) -> i64 {
     let lines: Vec<&str> = input.lines().collect();
     let o2_rating = find_rating(lines.clone(), true);
     let co2_rating = find_rating(lines, false);
